add AppNavigation & constants routes
