+++ App.tsx
@@ -1,21 +1,10 @@
-import 'react-native-gesture-handler';
 import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
 
-import Library from '@app/screens/library';
-import BookDetail from '@screens/book-detail';
-
-const Stack = createStackNavigator();
+import AppNavigation from './AppNavigation';
 
 const App = () => {
   return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen name="Library" component={Library} />
-        <Stack.Screen name="BookDetail" component={BookDetail} />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <AppNavigation />
   );
 };
