+++ src/app/screens/library/components/bookCard/index.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { Text, View, Image, TouchableOpacity } from 'react-native';
 import {useNavigation} from '@react-navigation/native';
+import { BOOK_DETAIL } from '@constants/routes';
+
 
 import { Book } from '@interfaces';
 
@@ -9,7 +11,7 @@ import styles from './styles/styles';
 const BookCard = ({ title, author, img }: Book) => {
   const navigation = useNavigation();
   return (
-    <TouchableOpacity onPress={() => navigation.navigate('BookDetail')} style={styles.button}>
+    <TouchableOpacity onPress={() => navigation.navigate(BOOK_DETAIL)} style={styles.button}>
       <View style={styles.container}>
         <Image source={img}/>
         <View style={styles.textContainer}>
