+++ src/constants/routes.js
@@ -1,0 +1,2 @@
+export const LIBRARY = 'Library';
+export const BOOK_DETAIL = 'bookDetail';
